Fix warning on 32 bit compile
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@2483 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/ppm/ppmtoarbtxt.c
@@ -471,7 +471,7 @@ validateFormatWithPpf(const char *       const format,
 
     default:
         pm_asprintf(errorP, "Has %lu extra transformation%s ",
-                    n-1, n-1 > 1 ? "s" : "");
+                    (unsigned long)n-1, n-1 > 1 ? "s" : "");
         break;
     }
 }
